move mysqlsupport to core
