+++ bukkit/src/main/java/io/github/wysohn/triggerreactor/bukkit/main/AbstractJavaPlugin.java
@@ -44,6 +44,7 @@ import io.github.wysohn.triggerreactor.core.main.TriggerReactorCore;
 import io.github.wysohn.triggerreactor.core.manager.Manager;
 import io.github.wysohn.triggerreactor.core.module.CorePluginModule;
 import io.github.wysohn.triggerreactor.tools.ContinuingTasks;
+import io.github.wysohn.triggerreactor.tools.mysql.MysqlSupport;
 import org.bukkit.Bukkit;
 import org.bukkit.Server;
 import org.bukkit.command.Command;
